# Fw 4943 more performance optimizations in the search api