--- a/firstvoices/backend/serializers/base_serializers.py
+++ b/firstvoices/backend/serializers/base_serializers.py
@@ -203,10 +203,3 @@ class WritableControlledSiteContentSerializer(
 
     class Meta(BaseControlledSiteContentSerializer.Meta):
         fields = BaseControlledSiteContentSerializer.Meta.fields
-
-
-class LinkedSiteMinimalSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Site
-        fields = ("id", "title", "slug")
-        read_only_fields = ("id", "title", "slug")