--- a/firstvoices/backend/serializers/media_serializers.py
+++ b/firstvoices/backend/serializers/media_serializers.py
@@ -262,3 +262,18 @@ class RelatedMediaSerializerMixin(metaclass=serializers.SerializerMetaclass):
 
     class Meta:
         fields = ("related_audio", "related_images", "related_videos")
+
+
+class RelatedAudioMinimalSerializer(serializers.ModelSerializer):
+    original = MediaFileUploadSerializer(read_only=True)
+
+    class Meta:
+        model = media.Audio
+        fields = ("id", "original")
+        read_only_fields = ("id", "original")
+
+
+class RelatedImageMinimalSerializer(RelatedAudioMinimalSerializer):
+    class Meta:
+        model = media.Image
+        fields = RelatedAudioMinimalSerializer.Meta.fields