### Description of Changes
- Reduced the first hit time from ~3 to under 2s.
- Reduced the consequtive hits times from ~1.5s to ~700ms~ ~800ms.
- Added minimal serializers for dictionary_entry, site, image and audio.
- Changed response structure to only include fields which are required to display the search results.

### Checklist
- ~README / documentation has been updated if needed~
- [x] PR title / commit messages contain Jira ticket number if applicable
- ~Tests have been updated / created if applicable~
- ~Admin Panel has been updated if models have been added or modified~
- ~Migrations have been updated and committed if applicable~
- ~Team Postman workspace has been updated if applicable~

### Reviewers Should Do The Following:
- [x] Review the code changes here
- ~Pull the branch and test locally~

### Additional Notes
N/A